move browser schemas into versioned folder
