+++ src/schemas/recording/index.ts
@@ -1,0 +1,1 @@
+export * from './v1/browser'
