+++ src/schemas/recording/v1/browser.ts
